Unreviewed, reverting 304590@main (ee5d5265c6da) https://bugs.webkit.org/show_bug.cgi?id=304341 rdar://166716256
REGRESSION(304590@main): Broke Internal builds

Reverted change:

    Re-land [Swift in WebKit] Fix the modularization of WebGPU
    https://bugs.webkit.org/show_bug.cgi?id=304307
    rdar://166669345
    304590@main (ee5d5265c6da)

Canonical link: https://commits.webkit.org/304609@main
