+++ Source/WTF/Configurations/WTF.xcconfig
@@ -64,7 +64,6 @@ GENERATE_WTF_MODULEMAP[sdk=macosx26.2*] = $(WK_NOT_$(DEPLOYMENT_LOCATION));
 GENERATE_WTF_MODULEMAP[sdk=macosx26.3*] = $(WK_NOT_$(DEPLOYMENT_LOCATION));
 
 GENERATE_WTF_MODULEMAP[sdk=iphone*] = $(GENERATE_WTF_MODULEMAP$(WK_IOS_26));
-GENERATE_WTF_MODULEMAP[sdk=xr*] = $(GENERATE_WTF_MODULEMAP$(WK_IOS_26));
 GENERATE_WTF_MODULEMAP_IOS_SINCE_26 = YES;
 
 GENERATE_WTF_MODULEMAP[sdk=bridgeos*] = ;
