+++ Source/WebGPU/Configurations/WebGPU.xcconfig
@@ -62,15 +62,6 @@ GCC_PREPROCESSOR_DEFINITIONS_ENABLE_WEBGPU_SWIFT = ENABLE_WEBGPU_SWIFT;
 
 DEFINES_MODULE = YES;
 MODULEMAP_FILE = WebGPU/WebGPU.modulemap;
-MODULEMAP_PRIVATE_FILE = WebGPU/WebGPU_Private.modulemap;
-ENABLE_MODULE_VERIFIER = $(WK_ENABLE_MODULE_VERIFIER_$(ENABLE_WEBGPU_SWIFT));
-WK_ENABLE_MODULE_VERIFIER_ = NO;
-WK_ENABLE_MODULE_VERIFIER_ENABLE_WEBGPU_SWIFT = YES;
-MODULE_VERIFIER_SUPPORTED_LANGUAGES = objective-c++;
-MODULE_VERIFIER_SUPPORTED_LANGUAGE_STANDARDS = gnu++23;
-MODULE_VERIFIER_KIND = builtin;
-OTHER_MODULE_VERIFIER_FLAGS = -- -I$(BUILT_PRODUCTS_DIR)$(WK_LIBRARY_HEADERS_FOLDER_PATH) -I$(WK_ADDITIONAL_SDKS)$(WK_LIBRARY_HEADERS_FOLDER_PATH);
-
 
 SUPPORTS_TEXT_BASED_API = $(SUPPORTS_TEXT_BASED_API_$(OVERRIDE_SUPPORTS_TEXT_BASED_API));
 SUPPORTS_TEXT_BASED_API_ = YES;
@@ -109,12 +100,12 @@ OTHER_LDFLAGS = $(inherited) $(UNEXPORT_SWIFT_CXX_LDFLAGS) $(SOURCE_VERSION_LDFL
 
 ENABLE_WEBGPU_SWIFT = ;
 ENABLE_WEBGPU_SWIFT[sdk=macosx*] = ENABLE_WEBGPU_SWIFT;
-ENABLE_WEBGPU_SWIFT[sdk=macosx*14*] = ;
-ENABLE_WEBGPU_SWIFT[sdk=macosx*15*] = ;
 ENABLE_WEBGPU_SWIFT[sdk=macosx*26.0*] = ;
 ENABLE_WEBGPU_SWIFT[sdk=macosx*26.1*] = ;
 ENABLE_WEBGPU_SWIFT[sdk=macosx*26.2*] = ;
 ENABLE_WEBGPU_SWIFT[sdk=macosx*26.3*] = ;
+ENABLE_WEBGPU_SWIFT[sdk=macosx*14*] = ;
+ENABLE_WEBGPU_SWIFT[sdk=macosx*15*] = ;
 ENABLE_WEBGPU_SWIFT[sdk=iphoneos*] = ENABLE_WEBGPU_SWIFT;
 // FIXME: Support WebGPU swift in iOS simulator builds once the underlying
 // modules issue is fixed (https://bugs.webkit.org/show_bug.cgi?id=300146).
@@ -128,7 +119,6 @@ SWIFT_VERSION = 6.0;
 SWIFT_INSTALL_OBJC_HEADER = NO;
 SWIFT_OBJC_INTEROP_MODE = $(WK_SWIFT_OBJC_INTEROP_MODE_$(ENABLE_WEBGPU_SWIFT))
 WK_SWIFT_OBJC_INTEROP_MODE_ENABLE_WEBGPU_SWIFT = objcxx;
-SWIFT_LIBRARY_LEVEL = spi;
 
 // FIXME: reenable this once rdar://154887575 lands
 // SWIFT_UPCOMING_FEATURE_MEMBER_IMPORT_VISIBILITY = YES
