+++ Source/WebGPU/WebGPU/Buffer.mm
@@ -36,9 +36,6 @@
 
 #if ENABLE(WEBGPU_SWIFT)
 #import "CxxBridging.h"
-#import <WebGPU/CxxBridgingPublic.h>
-#import <WebGPU/WGPUTextureImpl.h>
-#import <WebGPU/WebGPU.h>
 #import "WebGPUSwift-Generated.h"
 
 DEFINE_SWIFTCXX_THUNK(WebGPU::Buffer, copyFrom, void, const std::span<const uint8_t>, const size_t);
