+++ Source/WebGPU/WebGPU/CommandBuffer.h
@@ -25,7 +25,6 @@
 
 #pragma once
 
-#import <Metal/Metal.h>
 #import <wtf/FastMalloc.h>
 #import <wtf/Ref.h>
 #import <wtf/RefCountedAndCanMakeWeakPtr.h>
