+++ Source/WebGPU/WebGPU/CommandEncoder.mm
@@ -38,9 +38,6 @@
 #import "TextureOrTextureView.h"
 #if ENABLE(WEBGPU_SWIFT)
 #import "CxxBridging.h"
-#import <WebGPU/CxxBridgingPublic.h>
-#import <WebGPU/WGPUTextureImpl.h>
-#import <WebGPU/WebGPU.h>
 #import "WebGPUSwift-Generated.h"
 #endif
 #import <wtf/CheckedArithmetic.h>
