+++ Source/WebGPU/WebGPU/CommandEncoder.swift
@@ -24,36 +24,17 @@
  */
 
 import Metal
-internal import WebGPU_Internal.Buffer
-internal import WebGPU_Internal.CommandEncoder
-internal import WebGPU_Internal.CxxBridging
-internal import WebGPU_Internal.QuerySet
-import wtf.Cpp
-
+import WebGPU_Internal
 public typealias WTFString = String
 public typealias String = Swift.String
 
-// FIXME: Eventually all these "thunks" should be removed.
-// swift-format-ignore: AllPublicDeclarationsHaveDocumentation
 public func clearBuffer(
-    commandEncoder: WebGPU.CommandEncoder,
-    buffer: WebGPU.Buffer,
-    offset: UInt64,
-    size: inout UInt64
+    commandEncoder: WebGPU.CommandEncoder, buffer: WebGPU.Buffer, offset: UInt64, size: inout UInt64
 ) {
     commandEncoder.clearBuffer(buffer: buffer, offset: offset, size: &size)
 }
-
-// FIXME: Eventually all these "thunks" should be removed.
-// swift-format-ignore: AllPublicDeclarationsHaveDocumentation
-public func resolveQuerySet(
-    commandEncoder: WebGPU.CommandEncoder,
-    querySet: WebGPU.QuerySet,
-    firstQuery: UInt32,
-    queryCount: UInt32,
-    destination: WebGPU.Buffer,
-    destinationOffset: UInt64
-) {
+public func resolveQuerySet(commandEncoder: WebGPU.CommandEncoder, querySet: WebGPU.QuerySet, firstQuery: UInt32, queryCount:UInt32, destination: WebGPU.Buffer, destinationOffset: UInt64)
+{
     commandEncoder.resolveQuerySet(querySet, firstQuery: firstQuery, queryCount: queryCount, destination: destination, destinationOffset: destinationOffset)
 }
 
@@ -85,14 +66,7 @@ public func CommandEncoder_copyTextureToTexture_thunk(commandEncoder: WebGPU.Com
 // swift-format-ignore: AlwaysUseLowerCamelCase
 // swift-format-ignore: AllPublicDeclarationsHaveDocumentation
 @_expose(Cxx)
-public func CommandEncoder_copyBufferToBuffer_thunk(
-    commandEncoder: WebGPU.CommandEncoder,
-    source: WebGPU.Buffer,
-    sourceOffset: UInt64,
-    destination: WebGPU.Buffer,
-    destinationOffset: UInt64,
-    size: UInt64
-) {
+public func CommandEncoder_copyBufferToBuffer_thunk(commandEncoder: WebGPU.CommandEncoder, source: WebGPU.Buffer, sourceOffset: UInt64, destination: WebGPU.Buffer, destinationOffset: UInt64, size: UInt64) {
     commandEncoder.copyBufferToBuffer(source: source, sourceOffset: sourceOffset, destination: destination, destinationOffset: destinationOffset, size: size)
 }
 
@@ -282,21 +256,16 @@ extension WebGPU.CommandEncoder {
 
         return result
     }
-
     public func clearTextureIfNeeded(destination: WGPUImageCopyTexture, slice: UInt) {
-        WebGPU.CommandEncoder.clearTextureIfNeeded(destination, slice, m_device.ptr(), m_blitCommandEncoder)
+        return WebGPU.CommandEncoder.clearTextureIfNeeded(destination, slice, m_device.ptr(), m_blitCommandEncoder)
     }
-
-    private func clearTextureIfNeeded(
-        destination: WGPUImageCopyTexture,
-        slice: UInt,
-        device: WebGPU.Device,
-        blitCommandEncoder: MTLBlitCommandEncoder?
-    ) {
+    private func clearTextureIfNeeded(destination: WGPUImageCopyTexture , slice: UInt, device: WebGPU.Device , blitCommandEncoder: MTLBlitCommandEncoder?)
+    {
         let texture = WebGPU.fromAPI(destination.texture)
         let mipLevel: UInt = UInt(destination.mipLevel)
         clearTextureIfNeeded(texture, mipLevel, slice, device, blitCommandEncoder)
     }
+
 
     private func clearTextureIfNeeded(
         _ texture: WebGPU.Texture,
