+++ Source/WebGPU/WebGPU/CxxBridging.h
@@ -31,6 +31,9 @@
 #include <wtf/MathExtras.h>
 #include <wtf/Ref.h>
 
+using SpanConstUInt8 = std::span<const uint8_t>;
+using SpanUInt8 = std::span<uint8_t>;
+
 inline unsigned long roundUpToMultipleOfNonPowerOfTwoCheckedUInt32UnsignedLong(Checked<uint32_t> x, unsigned long y) { return WTF::roundUpToMultipleOfNonPowerOfTwo(x, y); }
 inline uint32_t roundUpToMultipleOfNonPowerOfTwoUInt32UInt32(uint32_t a, uint32_t b) { return WTF::roundUpToMultipleOfNonPowerOfTwo<uint32_t, Checked<uint32_t>>(a, b); }
