+++ Source/WebGPU/WebGPU/Internal/WebGPUSwiftInternal.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2023 Apple Inc. All rights reserved.
+ * Copyright (c) 2021-2025 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -25,10 +25,19 @@
 
 #pragma once
 
-#import <WebGPU/WebGPU.h>
-#import <wtf/SwiftBridging.h>
-
-// FIXME(rdar://155970441): this annotation should be in WebGPU.h, move it once we support
-// annotating incomplete types
-struct SWIFT_SHARED_REFERENCE(wgpuTextureReference, wgpuTextureRelease) WGPUTextureImpl {
-};
+#include "APIConversions.h"
+#include "Buffer.h"
+#include "CommandBuffer.h"
+#include "CommandEncoder.h"
+#include "CommandsMixin.h"
+#include "ComputePassEncoder.h"
+#include "CxxBridging.h"
+#include "DDModelTypes.h"
+#include "Device.h"
+#include "IsValidToUseWith.h"
+#include "QuerySet.h"
+#include "Queue.h"
+#include "RenderPassEncoder.h"
+#include "Texture.h"
+#include "TextureOrTextureView.h"
+#include "TextureView.h"
