+++ Source/WebGPU/WebGPU/Queue.mm
@@ -37,8 +37,6 @@
 #import "TextureView.h"
 #if ENABLE(WEBGPU_SWIFT)
 #import "CxxBridging.h"
-#import <WebGPU/CxxBridgingPublic.h>
-#import <WebGPU/WGPUTextureImpl.h>
 #import "WebGPUSwift-Generated.h"
 #endif
 #import <simd/simd.h>
