+++ Source/WebGPU/WebGPU/TextureView.h
@@ -26,7 +26,6 @@
 #pragma once
 
 #import "BindableResource.h"
-#import <WebGPU/WGPUTextureViewImpl.h>
 #import <wtf/FastMalloc.h>
 #import <wtf/Ref.h>
 #import <wtf/RefCountedAndCanMakeWeakPtr.h>
@@ -34,6 +33,11 @@
 #import <wtf/TZoneMalloc.h>
 #import <wtf/WeakHashSet.h>
 #import <wtf/WeakPtr.h>
+
+// FIXME(rdar://155970441): this annotation should be in WebGPU.h, move it once we support
+// annotating incomplete types
+struct SWIFT_SHARED_REFERENCE(wgpuTextureViewReference, wgpuTextureViewRelease) WGPUTextureViewImpl {
+};
 
 namespace WebGPU {
