+++ Source/WebGPU/WebGPU/WebGPU.h
@@ -30,6 +30,8 @@
 #ifndef WEBGPU_H_
 #define WEBGPU_H_
 
+#include <wtf/SwiftBridging.h>
+
 #if __has_include(<swift/bridging>)
 #include <swift/bridging>
 #endif
