+++ Source/WebGPU/WebGPU/WebGPU.modulemap
@@ -1,7 +1,6 @@
-framework module WebGPU [system] {
-    umbrella "Headers"
-
-    explicit module * {
-        export *
-    }
+framework module WebGPU {
+  umbrella header "WebGPU.h"
+  header "WebGPUExt.h"
+  export *
+  module * { export * }
 }
