--- Source/WebGPU/WebGPU/WebGPU_Private.modulemap
@@ -1,7 +1,0 @@
-framework module WebGPU_Private [system] {
-    umbrella "PrivateHeaders"
-
-    explicit module * {
-        export *
-    }
-}
